spliced out Address from Member class
