+++ Params.h
@@ -8,8 +8,6 @@
 #define _PARAMS_H_
 
 #include "stdincludes.h"
-#include "Params.h"
-#include "Member.h"
 
 enum testTYPE { CREATE_TEST, READ_TEST, UPDATE_TEST, DELETE_TEST };
 
@@ -18,23 +16,24 @@ enum testTYPE { CREATE_TEST, READ_TEST, UPDATE_TEST, DELETE_TEST };
  *
  * DESCRIPTION: Params class describing the test cases
  */
-class Params{
+class Params
+{
 public:
-	int MAX_NNB;                // max number of neighbors
-	int SINGLE_FAILURE;			// single/multi failure
-	double MSG_DROP_PROB;		// message drop probability
-	double STEP_RATE;		    // dictates the rate of insertion
-	int EN_GPSZ;			    // actual number of peers
-	int MAX_MSG_SIZE;
-	int DROP_MSG;
-	int dropmsg;
-	int globaltime;
-	int allNodesJoined;
-	short PORTNUM;
-	int CRUDTEST;
-	Params();
-	void setparams(char *);
-	int getcurrtime();
+    int MAX_NNB;                // max number of neighbors
+    int SINGLE_FAILURE;			// single/multi failure
+    double MSG_DROP_PROB;		// message drop probability
+    double STEP_RATE;		    // dictates the rate of insertion
+    int EN_GPSZ;			    // actual number of peers
+    int MAX_MSG_SIZE;
+    int DROP_MSG;
+    int dropmsg;
+    int globaltime;
+    int allNodesJoined;
+    short PORTNUM;
+    int CRUDTEST;
+    Params();
+    void setparams(char *);
+    int getcurrtime();
 };
 
 #endif /* _PARAMS_H_ */
